Stablize `non_ascii_idents` feature.
