+++ src/test/debuginfo/multi-byte-chars.rs
@@ -2,8 +2,6 @@
 
 // compile-flags:-g
 
-#![feature(non_ascii_idents)]
-
 // This test checks whether debuginfo generation can handle multi-byte UTF-8
 // characters at the end of a block. There's no need to do anything in the
 // debugger -- just make sure that the compiler doesn't crash.
