+++ src/test/ui/codemap_tests/unicode_2.rs
@@ -1,5 +1,3 @@
-#![feature(non_ascii_idents)]
-
 fn main() {
     let _ = ("a̐éö̲", 0u7); //~ ERROR invalid width
     let _ = ("아あ", 1i42); //~ ERROR invalid width
