+++ src/test/ui/codemap_tests/unicode_2.stderr
@@ -1,5 +1,5 @@
 error: invalid width `7` for integer literal
-  --> $DIR/unicode_2.rs:4:25
+  --> $DIR/unicode_2.rs:2:25
    |
 LL |     let _ = ("a̐éö̲", 0u7);
    |                     ^^^
@@ -7,7 +7,7 @@ LL |     let _ = ("a̐éö̲", 0u7);
    = help: valid widths are 8, 16, 32, 64 and 128
 
 error: invalid width `42` for integer literal
-  --> $DIR/unicode_2.rs:5:20
+  --> $DIR/unicode_2.rs:3:20
    |
 LL |     let _ = ("아あ", 1i42);
    |                      ^^^^
@@ -15,7 +15,7 @@ LL |     let _ = ("아あ", 1i42);
    = help: valid widths are 8, 16, 32, 64 and 128
 
 error[E0425]: cannot find value `a̐é` in this scope
-  --> $DIR/unicode_2.rs:6:13
+  --> $DIR/unicode_2.rs:4:13
    |
 LL |     let _ = a̐é;
    |             ^^ not found in this scope
