--- src/test/ui/feature-gates/feature-gate-non_ascii_idents.stderr
@@ -1,130 +1,0 @@
-error: items in `extern` blocks cannot use non-ascii identifiers
-  --> $DIR/feature-gate-non_ascii_idents.rs:30:8
-   |
-LL | extern "C" {
-   | ---------- in this `extern` block
-LL |     fn qüx();
-   |        ^^^
-   |
-   = note: This limitation may be lifted in the future; see issue #83942 <https://github.com/rust-lang/rust/issues/83942> for more information
-
-error[E0658]: non-ascii idents are not fully supported
-  --> $DIR/feature-gate-non_ascii_idents.rs:1:22
-   |
-LL | extern crate core as bäz;
-   |                      ^^^
-   |
-   = note: see issue #55467 <https://github.com/rust-lang/rust/issues/55467> for more information
-   = help: add `#![feature(non_ascii_idents)]` to the crate attributes to enable
-
-error[E0658]: non-ascii idents are not fully supported
-  --> $DIR/feature-gate-non_ascii_idents.rs:3:5
-   |
-LL | use föö::bar;
-   |     ^^^
-   |
-   = note: see issue #55467 <https://github.com/rust-lang/rust/issues/55467> for more information
-   = help: add `#![feature(non_ascii_idents)]` to the crate attributes to enable
-
-error[E0658]: non-ascii idents are not fully supported
-  --> $DIR/feature-gate-non_ascii_idents.rs:5:5
-   |
-LL | mod föö {
-   |     ^^^
-   |
-   = note: see issue #55467 <https://github.com/rust-lang/rust/issues/55467> for more information
-   = help: add `#![feature(non_ascii_idents)]` to the crate attributes to enable
-
-error[E0658]: non-ascii idents are not fully supported
-  --> $DIR/feature-gate-non_ascii_idents.rs:9:4
-   |
-LL | fn bär(
-   |    ^^^
-   |
-   = note: see issue #55467 <https://github.com/rust-lang/rust/issues/55467> for more information
-   = help: add `#![feature(non_ascii_idents)]` to the crate attributes to enable
-
-error[E0658]: non-ascii idents are not fully supported
-  --> $DIR/feature-gate-non_ascii_idents.rs:10:5
-   |
-LL |     bäz: isize
-   |     ^^^
-   |
-   = note: see issue #55467 <https://github.com/rust-lang/rust/issues/55467> for more information
-   = help: add `#![feature(non_ascii_idents)]` to the crate attributes to enable
-
-error[E0658]: non-ascii idents are not fully supported
-  --> $DIR/feature-gate-non_ascii_idents.rs:12:9
-   |
-LL |     let _ö: isize;
-   |         ^^
-   |
-   = note: see issue #55467 <https://github.com/rust-lang/rust/issues/55467> for more information
-   = help: add `#![feature(non_ascii_idents)]` to the crate attributes to enable
-
-error[E0658]: non-ascii idents are not fully supported
-  --> $DIR/feature-gate-non_ascii_idents.rs:15:10
-   |
-LL |         (_ä, _) => {}
-   |          ^^
-   |
-   = note: see issue #55467 <https://github.com/rust-lang/rust/issues/55467> for more information
-   = help: add `#![feature(non_ascii_idents)]` to the crate attributes to enable
-
-error[E0658]: non-ascii idents are not fully supported
-  --> $DIR/feature-gate-non_ascii_idents.rs:19:8
-   |
-LL | struct Föö {
-   |        ^^^
-   |
-   = note: see issue #55467 <https://github.com/rust-lang/rust/issues/55467> for more information
-   = help: add `#![feature(non_ascii_idents)]` to the crate attributes to enable
-
-error[E0658]: non-ascii idents are not fully supported
-  --> $DIR/feature-gate-non_ascii_idents.rs:20:5
-   |
-LL |     föö: isize
-   |     ^^^
-   |
-   = note: see issue #55467 <https://github.com/rust-lang/rust/issues/55467> for more information
-   = help: add `#![feature(non_ascii_idents)]` to the crate attributes to enable
-
-error[E0658]: non-ascii idents are not fully supported
-  --> $DIR/feature-gate-non_ascii_idents.rs:23:6
-   |
-LL | enum Bär {
-   |      ^^^
-   |
-   = note: see issue #55467 <https://github.com/rust-lang/rust/issues/55467> for more information
-   = help: add `#![feature(non_ascii_idents)]` to the crate attributes to enable
-
-error[E0658]: non-ascii idents are not fully supported
-  --> $DIR/feature-gate-non_ascii_idents.rs:24:5
-   |
-LL |     Bäz {
-   |     ^^^
-   |
-   = note: see issue #55467 <https://github.com/rust-lang/rust/issues/55467> for more information
-   = help: add `#![feature(non_ascii_idents)]` to the crate attributes to enable
-
-error[E0658]: non-ascii idents are not fully supported
-  --> $DIR/feature-gate-non_ascii_idents.rs:25:9
-   |
-LL |         qüx: isize
-   |         ^^^
-   |
-   = note: see issue #55467 <https://github.com/rust-lang/rust/issues/55467> for more information
-   = help: add `#![feature(non_ascii_idents)]` to the crate attributes to enable
-
-error[E0658]: non-ascii idents are not fully supported
-  --> $DIR/feature-gate-non_ascii_idents.rs:30:8
-   |
-LL |     fn qüx();
-   |        ^^^
-   |
-   = note: see issue #55467 <https://github.com/rust-lang/rust/issues/55467> for more information
-   = help: add `#![feature(non_ascii_idents)]` to the crate attributes to enable
-
-error: aborting due to 14 previous errors
-
-For more information about this error, try `rustc --explain E0658`.
