+++ src/test/ui/imports/local-modularized-tricky-fail-2.stderr
@@ -1,42 +1,43 @@
-error[E0658]: non-ascii idents are not fully supported
-  --> $DIR/local-modularized-tricky-fail-2.rs:20:32
+error: macro-expanded `macro_export` macros from the current crate cannot be referred to by absolute paths
+  --> $DIR/local-modularized-tricky-fail-2.rs:13:9
    |
-LL | exported!();
-   | ------------ in this macro invocation
+LL |     use exported;
+   |         ^^^^^^^^
+   |
+   = note: `#[deny(macro_expanded_macro_exports_accessed_by_absolute_paths)]` on by default
+   = warning: this was previously accepted by the compiler but is being phased out; it will become a hard error in a future release!
+   = note: for more information, see issue #52234 <https://github.com/rust-lang/rust/issues/52234>
+note: the macro is defined here
+  --> $DIR/local-modularized-tricky-fail-2.rs:5:5
+   |
+LL | /     macro_rules! exported {
+LL | |         () => ()
+LL | |     }
+   | |_____^
 ...
-LL |                 () => ( struct Б; )
-   |                                ^
-   |
-   = note: see issue #55467 <https://github.com/rust-lang/rust/issues/55467> for more information
-   = help: add `#![feature(non_ascii_idents)]` to the crate attributes to enable
+LL |   define_exported!();
+   |   ------------------- in this macro invocation
    = note: this error originates in a macro (in Nightly builds, run with -Z macro-backtrace for more info)
 
-error[E0658]: non-ascii idents are not fully supported
-  --> $DIR/local-modularized-tricky-fail-2.rs:36:24
+error: macro-expanded `macro_export` macros from the current crate cannot be referred to by absolute paths
+  --> $DIR/local-modularized-tricky-fail-2.rs:19:5
    |
-LL |     panic!();
-   |     --------- in this macro invocation
+LL |     ::exported!();
+   |     ^^^^^^^^^^
+   |
+   = warning: this was previously accepted by the compiler but is being phased out; it will become a hard error in a future release!
+   = note: for more information, see issue #52234 <https://github.com/rust-lang/rust/issues/52234>
+note: the macro is defined here
+  --> $DIR/local-modularized-tricky-fail-2.rs:5:5
+   |
+LL | /     macro_rules! exported {
+LL | |         () => ()
+LL | |     }
+   | |_____^
 ...
-LL |         () => ( struct Г; )
-   |                        ^
-   |
-   = note: see issue #55467 <https://github.com/rust-lang/rust/issues/55467> for more information
-   = help: add `#![feature(non_ascii_idents)]` to the crate attributes to enable
+LL |   define_exported!();
+   |   ------------------- in this macro invocation
    = note: this error originates in a macro (in Nightly builds, run with -Z macro-backtrace for more info)
 
-error[E0658]: non-ascii idents are not fully supported
-  --> $DIR/local-modularized-tricky-fail-2.rs:46:24
-   |
-LL | include!();
-   | ----------- in this macro invocation
-...
-LL |         () => ( struct Д; )
-   |                        ^
-   |
-   = note: see issue #55467 <https://github.com/rust-lang/rust/issues/55467> for more information
-   = help: add `#![feature(non_ascii_idents)]` to the crate attributes to enable
-   = note: this error originates in a macro (in Nightly builds, run with -Z macro-backtrace for more info)
+error: aborting due to 2 previous errors
 
-error: aborting due to 3 previous errors
-
-For more information about this error, try `rustc --explain E0658`.
