--- src/test/ui/imports/local-modularized-tricky-fail-3.rs
@@ -1,22 +1,0 @@
-// Crate-local macro expanded `macro_export` macros cannot be accessed with module-relative paths.
-
-macro_rules! define_exported { () => {
-    #[macro_export]
-    macro_rules! exported {
-        () => ()
-    }
-}}
-
-define_exported!();
-
-mod m {
-    use exported;
-    //~^ ERROR macro-expanded `macro_export` macros from the current crate cannot
-    //~| WARN this was previously accepted
-}
-
-fn main() {
-    ::exported!();
-    //~^ ERROR macro-expanded `macro_export` macros from the current crate cannot
-    //~| WARN this was previously accepted
-}
