+++ src/test/ui/imports/local-modularized-tricky-pass-1.rs
