+++ src/test/ui/issues/issue-44023.rs
@@ -1,5 +1,3 @@
-#![feature(non_ascii_idents)]
-
 pub fn main () {}
 
 fn საჭმელად_გემრიელი_სადილი ( ) -> isize { //~ ERROR mismatched types
