+++ src/test/ui/issues/issue-44023.stderr
@@ -1,5 +1,5 @@
 error[E0308]: mismatched types
-  --> $DIR/issue-44023.rs:5:36
+  --> $DIR/issue-44023.rs:3:36
    |
 LL | fn საჭმელად_გემრიელი_სადილი ( ) -> isize {
    |    ------------------------        ^^^^^ expected `isize`, found `()`
