+++ src/test/ui/lint/lint-non-snake-case-no-lowercase-equivalent.rs
@@ -3,7 +3,6 @@
 #![allow(dead_code)]
 // pretty-expanded FIXME #23616
 
-#![feature(non_ascii_idents)]
 #![deny(non_snake_case)]
 
 // This name is neither upper nor lower case
