+++ src/test/ui/lint/lint-nonstandard-style-unicode-1.stderr
@@ -1,5 +1,5 @@
 error: type `χa` should have an upper camel case name
-  --> $DIR/lint-nonstandard-style-unicode-1.rs:15:8
+  --> $DIR/lint-nonstandard-style-unicode-1.rs:14:8
    |
 LL | struct χa;
    |        ^^ help: convert the identifier to upper camel case: `Χa`
@@ -11,37 +11,37 @@ LL | #![forbid(non_camel_case_types)]
    |           ^^^^^^^^^^^^^^^^^^^^
 
 error: type `__χa` should have an upper camel case name
-  --> $DIR/lint-nonstandard-style-unicode-1.rs:23:8
+  --> $DIR/lint-nonstandard-style-unicode-1.rs:22:8
    |
 LL | struct __χa;
    |        ^^^^ help: convert the identifier to upper camel case: `Χa`
 
 error: type `对__否` should have an upper camel case name
-  --> $DIR/lint-nonstandard-style-unicode-1.rs:28:8
+  --> $DIR/lint-nonstandard-style-unicode-1.rs:27:8
    |
 LL | struct 对__否;
    |        ^^^^^^ help: convert the identifier to upper camel case: `对_否`
 
 error: type `ヒ__χ` should have an upper camel case name
-  --> $DIR/lint-nonstandard-style-unicode-1.rs:31:8
+  --> $DIR/lint-nonstandard-style-unicode-1.rs:30:8
    |
 LL | struct ヒ__χ;
    |        ^^^^^ help: convert the identifier to upper camel case: `ヒΧ`
 
 error: type `Hello_你好` should have an upper camel case name
-  --> $DIR/lint-nonstandard-style-unicode-1.rs:37:8
+  --> $DIR/lint-nonstandard-style-unicode-1.rs:36:8
    |
 LL | struct Hello_你好;
    |        ^^^^^^^^^^ help: convert the identifier to upper camel case: `Hello你好`
 
 error: type `Hello_World` should have an upper camel case name
-  --> $DIR/lint-nonstandard-style-unicode-1.rs:40:8
+  --> $DIR/lint-nonstandard-style-unicode-1.rs:39:8
    |
 LL | struct Hello_World;
    |        ^^^^^^^^^^^ help: convert the identifier to upper camel case: `HelloWorld`
 
 error: type `你_ӟ` should have an upper camel case name
-  --> $DIR/lint-nonstandard-style-unicode-1.rs:43:8
+  --> $DIR/lint-nonstandard-style-unicode-1.rs:42:8
    |
 LL | struct 你_ӟ;
    |        ^^^^ help: convert the identifier to upper camel case: `你Ӟ`
