+++ src/test/ui/lint/lint-nonstandard-style-unicode-3.rs
@@ -1,7 +1,6 @@
 #![allow(dead_code)]
 
 #![forbid(non_upper_case_globals)]
-#![feature(non_ascii_idents)]
 
 // Some scripts (e.g., hiragana) don't have a concept of
 // upper/lowercase
