+++ src/test/ui/lint/rfc-2457-non-ascii-idents/lint-confusable-idents.stderr
@@ -1,5 +1,5 @@
 error: identifier pair considered confusable between `ｓ` and `s`
-  --> $DIR/lint-confusable-idents.rs:9:9
+  --> $DIR/lint-confusable-idents.rs:8:9
    |
 LL | const ｓ: usize = 42;
    |       -- this is where the previous identifier occurred
@@ -8,13 +8,13 @@ LL |     let s = "rust";
    |         ^
    |
 note: the lint level is defined here
-  --> $DIR/lint-confusable-idents.rs:2:9
+  --> $DIR/lint-confusable-idents.rs:1:9
    |
 LL | #![deny(confusable_idents)]
    |         ^^^^^^^^^^^^^^^^^
 
 error: identifier pair considered confusable between `s_s` and `ｓ_ｓ`
-  --> $DIR/lint-confusable-idents.rs:10:9
+  --> $DIR/lint-confusable-idents.rs:9:9
    |
 LL | const s_s: usize = 42;
    |       --- this is where the previous identifier occurred
