+++ src/test/ui/lint/rfc-2457-non-ascii-idents/lint-mixed-script-confusables-2.rs
@@ -1,5 +1,4 @@
 // check-pass
-#![feature(non_ascii_idents)]
 #![deny(mixed_script_confusables)]
 
 struct ΑctuallyNotLatin;
