+++ src/test/ui/lint/rfc-2457-non-ascii-idents/lint-mixed-script-confusables.stderr
@@ -1,11 +1,11 @@
 error: The usage of Script Group `Greek` in this crate consists solely of mixed script confusables
-  --> $DIR/lint-mixed-script-confusables.rs:4:8
+  --> $DIR/lint-mixed-script-confusables.rs:3:8
    |
 LL | struct ΑctuallyNotLatin;
    |        ^^^^^^^^^^^^^^^^
    |
 note: the lint level is defined here
-  --> $DIR/lint-mixed-script-confusables.rs:2:9
+  --> $DIR/lint-mixed-script-confusables.rs:1:9
    |
 LL | #![deny(mixed_script_confusables)]
    |         ^^^^^^^^^^^^^^^^^^^^^^^^
@@ -13,7 +13,7 @@ LL | #![deny(mixed_script_confusables)]
    = note: Please recheck to make sure their usages are indeed what you want.
 
 error: The usage of Script Group `Cyrillic` in this crate consists solely of mixed script confusables
-  --> $DIR/lint-mixed-script-confusables.rs:11:5
+  --> $DIR/lint-mixed-script-confusables.rs:10:5
    |
 LL | mod роре {
    |     ^^^^
@@ -22,7 +22,7 @@ LL | mod роре {
    = note: Please recheck to make sure their usages are indeed what you want.
 
 error: The usage of Script Group `Japanese, Katakana` in this crate consists solely of mixed script confusables
-  --> $DIR/lint-mixed-script-confusables.rs:13:11
+  --> $DIR/lint-mixed-script-confusables.rs:12:11
    |
 LL |     const エ: &'static str = "アイウ";
    |           ^^
