+++ src/test/ui/lint/rfc-2457-non-ascii-idents/lint-non-ascii-idents.stderr
@@ -1,23 +1,23 @@
 error: identifier contains non-ASCII characters
-  --> $DIR/lint-non-ascii-idents.rs:4:7
+  --> $DIR/lint-non-ascii-idents.rs:3:7
    |
 LL | const חלודה: usize = 2;
    |       ^^^^^
    |
 note: the lint level is defined here
-  --> $DIR/lint-non-ascii-idents.rs:2:9
+  --> $DIR/lint-non-ascii-idents.rs:1:9
    |
 LL | #![deny(non_ascii_idents)]
    |         ^^^^^^^^^^^^^^^^
 
 error: identifier contains non-ASCII characters
-  --> $DIR/lint-non-ascii-idents.rs:6:4
+  --> $DIR/lint-non-ascii-idents.rs:5:4
    |
 LL | fn coöperation() {}
    |    ^^^^^^^^^^^
 
 error: identifier contains non-ASCII characters
-  --> $DIR/lint-non-ascii-idents.rs:9:9
+  --> $DIR/lint-non-ascii-idents.rs:8:9
    |
 LL |     let naïveté = 2;
    |         ^^^^^^^
