+++ src/test/ui/lint/rfc-2457-non-ascii-idents/lint-uncommon-codepoints.stderr
@@ -1,23 +1,23 @@
 error: identifier contains uncommon Unicode codepoints
-  --> $DIR/lint-uncommon-codepoints.rs:4:7
+  --> $DIR/lint-uncommon-codepoints.rs:3:7
    |
 LL | const µ: f64 = 0.000001;
    |       ^
    |
 note: the lint level is defined here
-  --> $DIR/lint-uncommon-codepoints.rs:2:9
+  --> $DIR/lint-uncommon-codepoints.rs:1:9
    |
 LL | #![deny(uncommon_codepoints)]
    |         ^^^^^^^^^^^^^^^^^^^
 
 error: identifier contains uncommon Unicode codepoints
-  --> $DIR/lint-uncommon-codepoints.rs:6:4
+  --> $DIR/lint-uncommon-codepoints.rs:5:4
    |
 LL | fn dĳkstra() {}
    |    ^^^^^^^
 
 error: identifier contains uncommon Unicode codepoints
-  --> $DIR/lint-uncommon-codepoints.rs:9:9
+  --> $DIR/lint-uncommon-codepoints.rs:8:9
    |
 LL |     let ㇻㇲㇳ = "rust";
    |         ^^^^^^
