+++ src/test/ui/parser/issue-48508.rs
@@ -10,7 +10,6 @@
 // ignore-pretty issue #37195
 // ignore-asmjs wasm2js does not support source maps yet
 
-#![feature(non_ascii_idents)]
 #![allow(uncommon_codepoints)]
 
 #[path = "issue-48508-aux.rs"]
