+++ src/test/ui/parser/multibyte-char-use-seperator-issue-80134.stderr
@@ -8,7 +8,7 @@ LL |     (()é);
    |        help: missing `,`
 
 error: expected one of `)`, `,`, `.`, `?`, or an operator, found `氷`
-  --> $DIR/multibyte-char-use-seperator-issue-80134.rs:8:8
+  --> $DIR/multibyte-char-use-seperator-issue-80134.rs:7:8
    |
 LL |     (()氷);
    |        -^
@@ -23,30 +23,11 @@ LL |     (()é);
    |        ^ not found in this scope
 
 error[E0425]: cannot find value `氷` in this scope
-  --> $DIR/multibyte-char-use-seperator-issue-80134.rs:8:8
+  --> $DIR/multibyte-char-use-seperator-issue-80134.rs:7:8
    |
 LL |     (()氷);
    |        ^^ not found in this scope
 
-error[E0658]: non-ascii idents are not fully supported
-  --> $DIR/multibyte-char-use-seperator-issue-80134.rs:4:8
-   |
-LL |     (()é);
-   |        ^
-   |
-   = note: see issue #55467 <https://github.com/rust-lang/rust/issues/55467> for more information
-   = help: add `#![feature(non_ascii_idents)]` to the crate attributes to enable
+error: aborting due to 4 previous errors
 
-error[E0658]: non-ascii idents are not fully supported
-  --> $DIR/multibyte-char-use-seperator-issue-80134.rs:8:8
-   |
-LL |     (()氷);
-   |        ^^
-   |
-   = note: see issue #55467 <https://github.com/rust-lang/rust/issues/55467> for more information
-   = help: add `#![feature(non_ascii_idents)]` to the crate attributes to enable
-
-error: aborting due to 6 previous errors
-
-Some errors have detailed explanations: E0425, E0658.
-For more information about an error, try `rustc --explain E0425`.
+For more information about this error, try `rustc --explain E0425`.
