+++ src/test/ui/rfc-2457/crate_name_nonascii_forbidden-1.rs
@@ -1,5 +1,3 @@
-#![feature(non_ascii_idents)]
-
 extern crate ьаг; //~ ERROR cannot load a crate with a non-ascii name `ьаг`
 
 fn main() {}
