+++ src/test/ui/rfc-2457/crate_name_nonascii_forbidden-1.stderr
@@ -1,5 +1,5 @@
 error: cannot load a crate with a non-ascii name `ьаг`
-  --> $DIR/crate_name_nonascii_forbidden-1.rs:3:1
+  --> $DIR/crate_name_nonascii_forbidden-1.rs:1:1
    |
 LL | extern crate ьаг;
    | ^^^^^^^^^^^^^^^^^
