+++ src/test/ui/rfc-2457/crate_name_nonascii_forbidden-2.rs
@@ -1,6 +1,5 @@
 // compile-flags:--extern му_сгате
 // edition:2018
-#![feature(non_ascii_idents)]
 
 use му_сгате::baz; //~  ERROR cannot load a crate with a non-ascii name `му_сгате`
