+++ src/test/ui/rfc-2457/crate_name_nonascii_forbidden-2.stderr
@@ -1,5 +1,5 @@
 error: cannot load a crate with a non-ascii name `му_сгате`
-  --> $DIR/crate_name_nonascii_forbidden-2.rs:5:5
+  --> $DIR/crate_name_nonascii_forbidden-2.rs:4:5
    |
 LL | use му_сгате::baz;
    |     ^^^^^^^^
