+++ src/test/ui/rfc-2457/extern_block_nonascii_forbidden.stderr
@@ -1,5 +1,5 @@
 error: items in `extern` blocks cannot use non-ascii identifiers
-  --> $DIR/extern_block_nonascii_forbidden.rs:5:10
+  --> $DIR/extern_block_nonascii_forbidden.rs:4:10
    |
 LL | extern "C" {
    | ---------- in this `extern` block
@@ -9,7 +9,7 @@ LL |     type 一;
    = note: This limitation may be lifted in the future; see issue #83942 <https://github.com/rust-lang/rust/issues/83942> for more information
 
 error: items in `extern` blocks cannot use non-ascii identifiers
-  --> $DIR/extern_block_nonascii_forbidden.rs:6:8
+  --> $DIR/extern_block_nonascii_forbidden.rs:5:8
    |
 LL | extern "C" {
    | ---------- in this `extern` block
@@ -20,7 +20,7 @@ LL |     fn 二();
    = note: This limitation may be lifted in the future; see issue #83942 <https://github.com/rust-lang/rust/issues/83942> for more information
 
 error: items in `extern` blocks cannot use non-ascii identifiers
-  --> $DIR/extern_block_nonascii_forbidden.rs:7:12
+  --> $DIR/extern_block_nonascii_forbidden.rs:6:12
    |
 LL | extern "C" {
    | ---------- in this `extern` block
