+++ src/test/ui/rfc-2457/mod_file_nonascii_forbidden.rs
@@ -1,5 +1,3 @@
-#![feature(non_ascii_idents)]
-
 mod řųśť; //~ trying to load file for
 //~^ file not found for
