+++ src/test/ui/rfc-2457/mod_file_nonascii_forbidden.stderr
@@ -1,5 +1,5 @@
 error[E0583]: file not found for module `řųśť`
-  --> $DIR/mod_file_nonascii_forbidden.rs:3:1
+  --> $DIR/mod_file_nonascii_forbidden.rs:1:1
    |
 LL | mod řųśť;
    | ^^^^^^^^^
@@ -7,7 +7,7 @@ LL | mod řųśť;
    = help: to create the module `řųśť`, create file "$DIR/řųśť.rs"
 
 error[E0754]: trying to load file for module `řųśť` with non-ascii identifier name
-  --> $DIR/mod_file_nonascii_forbidden.rs:3:5
+  --> $DIR/mod_file_nonascii_forbidden.rs:1:5
    |
 LL | mod řųśť;
    |     ^^^^
