+++ src/test/ui/rfc-2457/mod_file_nonascii_with_path_allowed.rs
@@ -1,5 +1,4 @@
 // check-pass
-#![feature(non_ascii_idents)]
 
 #[path="auxiliary/mod_file_nonascii_with_path_allowed-aux.rs"]
 mod řųśť;
