+++ src/test/ui/rfc-2457/mod_inline_nonascii_allowed.rs
@@ -1,5 +1,4 @@
 // check-pass
-#![feature(non_ascii_idents)]
 
 mod řųśť {
     const IS_GREAT: bool = true;
