+++ src/test/ui/rfc-2457/no_mangle_nonascii_forbidden.rs
@@ -1,5 +1,3 @@
-#![feature(non_ascii_idents)]
-
 #[no_mangle]
 pub fn řųśť() {}  //~ `#[no_mangle]` requires ASCII identifier
