--- src/test/ui/utf8_idents.stderr
@@ -1,47 +1,0 @@
-error[E0658]: non-ascii idents are not fully supported
-  --> $DIR/utf8_idents.rs:4:5
-   |
-LL |     'β,
-   |     ^^
-   |
-   = note: see issue #55467 <https://github.com/rust-lang/rust/issues/55467> for more information
-   = help: add `#![feature(non_ascii_idents)]` to the crate attributes to enable
-
-error[E0658]: non-ascii idents are not fully supported
-  --> $DIR/utf8_idents.rs:5:5
-   |
-LL |     γ
-   |     ^
-   |
-   = note: see issue #55467 <https://github.com/rust-lang/rust/issues/55467> for more information
-   = help: add `#![feature(non_ascii_idents)]` to the crate attributes to enable
-
-error[E0658]: non-ascii idents are not fully supported
-  --> $DIR/utf8_idents.rs:10:5
-   |
-LL |     δ: usize
-   |     ^
-   |
-   = note: see issue #55467 <https://github.com/rust-lang/rust/issues/55467> for more information
-   = help: add `#![feature(non_ascii_idents)]` to the crate attributes to enable
-
-error[E0658]: non-ascii idents are not fully supported
-  --> $DIR/utf8_idents.rs:14:9
-   |
-LL |     let α = 0.00001f64;
-   |         ^
-   |
-   = note: see issue #55467 <https://github.com/rust-lang/rust/issues/55467> for more information
-   = help: add `#![feature(non_ascii_idents)]` to the crate attributes to enable
-
-warning: type parameter `γ` should have an upper camel case name
-  --> $DIR/utf8_idents.rs:5:5
-   |
-LL |     γ
-   |     ^ help: convert the identifier to upper camel case: `Γ`
-   |
-   = note: `#[warn(non_camel_case_types)]` on by default
-
-error: aborting due to 4 previous errors; 1 warning emitted
-
-For more information about this error, try `rustc --explain E0658`.
